:100: Add test_amazon.py to test Amazon API
