+++ __init__.py
@@ -8,8 +8,10 @@ load_dotenv(verbose=True)
 
 app = Flask(__name__)
 
-amazon_api_client = AmazonAPI(os.getenv("amazon_access_key"), os.getenv("amazon_secret_key"),
-                              os.getenv("amazon_partner_tag"), "JP")
+amazon_api_client = AmazonAPI(os.getenv("amazon_access_key"),
+                              os.getenv("amazon_secret_key"),
+                              os.getenv("amazon_partner_tag"),
+                              "JP")
 
 if os.getenv("GAE_ENV", "").startswith("standard"):
     """ Production in GAE """
